features =)

+++ src/driver/eeprom.c
@@ -26,7 +26,7 @@ void EEPROM_WriteBuffer(uint32_t address, void *pBuffer, uint16_t size) {
   if (pBuffer == NULL) {
     return;
   }
-  const uint8_t PAGE_SIZE = SETTINGS_GetPageSize();
+  const uint16_t PAGE_SIZE = SETTINGS_GetPageSize();
 
   while (size) {
     uint16_t i = address % PAGE_SIZE;
@@ -56,7 +56,7 @@ void EEPROM_WriteBuffer(uint32_t address, void *pBuffer, uint16_t size) {
 }
 
 void EEPROM_ClearPage(uint16_t page) {
-  const uint8_t PAGE_SIZE = SETTINGS_GetPageSize();
+  const uint16_t PAGE_SIZE = SETTINGS_GetPageSize();
   const uint32_t address = page * PAGE_SIZE;
 
   uint8_t IIC_ADD = (uint8_t)(0xA0 | ((address / 0x10000) << 1));
