+++ src/driver/i2c.h
@@ -15,7 +15,7 @@ void I2C_Stop(void);
 uint8_t I2C_Read(bool bFinal);
 int I2C_Write(uint8_t Data);
 
-int I2C_ReadBuffer(void *pBuffer, uint8_t Size);
-int I2C_WriteBuffer(const void *pBuffer, uint8_t Size);
+uint16_t I2C_ReadBuffer(void *pBuffer, uint16_t Size);
+uint16_t I2C_WriteBuffer(const void *pBuffer, uint16_t Size);
 
 #endif
